skin-cancer-detection/requirements.txt
# Skin Cancer Detection - Requirements

# Web Framework
Flask==2.3.0
Werkzeug==2.3.0

# Deep Learning
tensorflow==2.13.0
keras==2.13.1

# Image Processing
Pillow==10.0.0
opencv-python==4.8.0.74

# Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0

# Development
jupyter==1.0.0
notebook==7.0.0
ipykernel==6.25.0
